tasmota-master/rootfs/app/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
paho-mqtt==1.6.1
zeroconf==0.127.0
requests==2.31.0
SQLAlchemy==2.0.23
python-socketio==5.10.0
eventlet==0.33.3
pyserial==3.5
websockets==12.0
python-dotenv==1.0.0
Werkzeug==3.0.1
aiohttp==3.9.1
APScheduler==3.10.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
python-magic==0.4.27
schedule==1.2.0
cryptography==41.0.7
bcrypt==4.1.2
psutil==5.9.6